Moved ReduxTextField and added option for passing style obj
