+++ client/src/components/ReduxTextField/index.js
@@ -2,12 +2,12 @@ import React from "react";
 import TextField from "@material-ui/core/TextField";
 
 const ReduxTextField = (props) => {
-  const { type, label, error, input } = props;
+  const { type, label, error, style, input } = props;
 
   return (
     <TextField
       label={label}
-      style={{ width: "60%", paddingBottom: "1.5rem" }}
+      style={style}
       type={type ? type : "text"}
       error={!!error}
       {...input}
